# contracts/openapi/tenant.v1.yaml
openapi: 3.1.0
info:
  title: AuraEDU Tenant Service API
  version: 1.0.0
  description: >
    Schools (tenants), branding, settings, and the feature-flag store.
    Source of truth for feature flags (spec §3). Owned by lane L1 (EP-05).
servers:
  - url: /api/v1
tags:
  - name: tenants
  - name: features

paths:
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List tenants (platform super admin only)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Tenant' } }
                  next_cursor: { type: [string, 'null'] }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant (onboarding / sign-up funnel)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantCreate' }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } }
        '422': { $ref: '#/components/responses/ValidationError' }

  /tenants/{code}:
    parameters:
      - $ref: '#/components/parameters/TenantCode'
    get:
      operationId: getTenant
      tags: [tenants]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Not found }
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant (platform super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantUpdate' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Not found }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant (platform super admin only)
      responses:
        '204': { description: No content }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Not found }

  /tenants/{code}/branding:
    parameters:
      - $ref: '#/components/parameters/TenantCode'
    get:
      operationId: getBranding
      tags: [tenants]
      summary: Public branding for a tenant
      security: []
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Branding' } } }
        '404': { description: Not found }

  /tenants/{code}/settings:
    parameters:
      - $ref: '#/components/parameters/TenantCode'
    get:
      operationId: getSettings
      tags: [tenants]
      summary: Tenant operational settings
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Not found }
    patch:
      operationId: updateSettings
      tags: [tenants]
      summary: Update tenant operational settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Settings' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Not found }
        '422': { $ref: '#/components/responses/ValidationError' }

  /tenants/resolve:
    get:
      operationId: resolveTenant
      tags: [tenants]
      summary: Resolve tenant by domain or subdomain
      security: []
      parameters:
        - name: domain
          in: query
          schema: { type: string }
        - name: subdomain
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Tenant' } } }
        '404': { description: Not found }

  /features:
    get:
      operationId: getFeatureSnapshot
      tags: [features]
      summary: Feature-flag snapshot for the resolved tenant (loaded by web/mobile at startup)
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [tenant_code, features]
                properties:
                  tenant_code: { type: string, example: upshs }
                  features:
                    type: array
                    items: { $ref: '#/components/schemas/FeatureFlag' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /features/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema: { type: string }
    put:
      operationId: setFeature
      tags: [features]
      summary: Enable or disable a feature for the resolved tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_enabled]
              properties:
                tenant_code: { type: string }
                is_enabled: { type: boolean }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/FeatureFlag' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Not found }

  /super-admin/features/{key}/override:
    parameters:
      - name: key
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: superAdminFeatureOverride
      tags: [features]
      summary: Platform super admin override for a tenant feature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_code, is_enabled]
              properties:
                tenant_code: { type: string }
                is_enabled: { type: boolean }
                reason: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/FeatureFlag' } } }
        '403': { $ref: '#/components/responses/Forbidden' }

components:
  parameters:
    TenantCode:
      name: code
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9-]{2,50}$' }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }

  schemas:
    Tenant:
      type: object
      required: [tenant_code, name, status]
      properties:
        tenant_code: { type: string, example: upshs }
        name: { type: string }
        short: { type: string }
        status: { type: string, enum: [active, suspended, onboarding] }
        domain: { type: [string, 'null'] }
        plan: { type: string, enum: [starter, growth, professional, ai_plus, enterprise] }
        branding: { $ref: '#/components/schemas/Branding' }
    TenantCreate:
      type: object
      required: [tenant_code, name]
      properties:
        tenant_code: { type: string, pattern: '^[a-z0-9-]{2,50}$' }
        name: { type: string }
        short: { type: string }
        status: { type: string, enum: [active, suspended, onboarding] }
        domain: { type: [string, 'null'] }
        plan: { type: string, enum: [starter, growth, professional, ai_plus, enterprise] }
        branding: { $ref: '#/components/schemas/Branding' }
    TenantUpdate:
      type: object
      properties:
        name: { type: string }
        short: { type: [string, 'null'] }
        status: { type: string, enum: [active, suspended, onboarding] }
        domain: { type: [string, 'null'] }
        plan: { type: string, enum: [starter, growth, professional, ai_plus, enterprise] }
        branding: { $ref: '#/components/schemas/Branding' }
    Branding:
      type: object
      properties:
        logo_url: { type: [string, 'null'] }
        brand:
          type: object
          properties:
            primary: { type: string, example: '#7B1113' }
            secondary: { type: [string, 'null'] }
    Settings:
      type: object
      properties:
        locale: { type: string, example: en-GH }
        timezone: { type: string, example: Africa/Accra }
        date_format: { type: string, example: DD/MM/YYYY }
        academic_year_start_month: { type: integer, minimum: 1, maximum: 12 }
        primary_contact_email: { type: [string, 'null'], format: email }
    FeatureFlag:
      type: object
      required: [feature_key, is_enabled]
      properties:
        feature_key: { type: string, example: ai_recommendations }
        is_enabled: { type: boolean }
        plan_required: { type: [string, 'null'] }
        config: { type: [object, 'null'], additionalProperties: true }
        rollout: { $ref: '#/components/schemas/RolloutConfig' }
    RolloutConfig:
      type: object
      properties:
        percentage: { type: integer, minimum: 0, maximum: 100 }
        updated_by: { type: string }
        reason: { type: string }

    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }

  responses:
    Forbidden:
      description: Not permitted (auth, tenant scope, RBAC, or feature disabled)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            feature_disabled: { value: { code: feature_disabled, message: "Feature not enabled for tenant" } }
            forbidden: { value: { code: forbidden, message: "Missing required permission" } }
            tenant_mismatch: { value: { code: tenant_mismatch, message: "Resource belongs to another tenant" } }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: validation_error, message: "tenant_code is invalid" }

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []