# packages/shared-types/gen/go/tenant/openapi.gen.go
// Code generated by @auraedu/codegen from the service OpenAPI contract. DO NOT EDIT.

package tenant

import (
	"context"
	"net/http"
)

// Tenant generated from OpenAPI schema.
type Tenant struct {
	TenantCode string    `json:"tenant_code"`
	Name       string    `json:"name"`
	Short      *string   `json:"short,omitempty"`
	Status     string    `json:"status"`
	Domain     *string   `json:"domain,omitempty"`
	Plan       *string   `json:"plan,omitempty"`
	Branding   *Branding `json:"branding,omitempty"`
}

// TenantCreate generated from OpenAPI schema.
type TenantCreate struct {
	TenantCode string    `json:"tenant_code"`
	Name       string    `json:"name"`
	Short      *string   `json:"short,omitempty"`
	Status     *string   `json:"status,omitempty"`
	Domain     *string   `json:"domain,omitempty"`
	Plan       *string   `json:"plan,omitempty"`
	Branding   *Branding `json:"branding,omitempty"`
}

// TenantUpdate generated from OpenAPI schema.
type TenantUpdate struct {
	Name     *string   `json:"name,omitempty"`
	Short    *string   `json:"short,omitempty"`
	Status   *string   `json:"status,omitempty"`
	Domain   *string   `json:"domain,omitempty"`
	Plan     *string   `json:"plan,omitempty"`
	Branding *Branding `json:"branding,omitempty"`
}

// Branding generated from OpenAPI schema.
type Branding struct {
	LogoUrl *string `json:"logo_url,omitempty"`
	Brand   *struct {
		Primary   *string `json:"primary,omitempty"`
		Secondary *string `json:"secondary,omitempty"`
	} `json:"brand,omitempty"`
}

// Settings generated from OpenAPI schema.
type Settings struct {
	Locale                 *string `json:"locale,omitempty"`
	Timezone               *string `json:"timezone,omitempty"`
	DateFormat             *string `json:"date_format,omitempty"`
	AcademicYearStartMonth *int    `json:"academic_year_start_month,omitempty"`
	PrimaryContactEmail    *string `json:"primary_contact_email,omitempty"`
}

// FeatureFlag generated from OpenAPI schema.
type FeatureFlag struct {
	FeatureKey   string                  `json:"feature_key"`
	IsEnabled    bool                    `json:"is_enabled"`
	PlanRequired *string                 `json:"plan_required,omitempty"`
	Config       *map[string]interface{} `json:"config,omitempty"`
	Rollout      *RolloutConfig          `json:"rollout,omitempty"`
}

// RolloutConfig generated from OpenAPI schema.
type RolloutConfig struct {
	Percentage *int    `json:"percentage,omitempty"`
	UpdatedBy  *string `json:"updated_by,omitempty"`
	Reason     *string `json:"reason,omitempty"`
}

// ServerInterface is implemented by the service HTTP adapter.
type ServerInterface interface {
	listTenants(w http.ResponseWriter, r *http.Request)
	createTenant(w http.ResponseWriter, r *http.Request)
	getTenant(w http.ResponseWriter, r *http.Request)
	updateTenant(w http.ResponseWriter, r *http.Request)
	deleteTenant(w http.ResponseWriter, r *http.Request)
	getBranding(w http.ResponseWriter, r *http.Request)
	getSettings(w http.ResponseWriter, r *http.Request)
	updateSettings(w http.ResponseWriter, r *http.Request)
	resolveTenant(w http.ResponseWriter, r *http.Request)
	getFeatureSnapshot(w http.ResponseWriter, r *http.Request)
	setFeature(w http.ResponseWriter, r *http.Request)
	superAdminFeatureOverride(w http.ResponseWriter, r *http.Request)
}

// ClientInterface is the generated consumer stub for this service.
type ClientInterface interface {
	listTenants(ctx context.Context) (*http.Response, error)
	createTenant(ctx context.Context) (*http.Response, error)
	getTenant(ctx context.Context) (*http.Response, error)
	updateTenant(ctx context.Context) (*http.Response, error)
	deleteTenant(ctx context.Context) (*http.Response, error)
	getBranding(ctx context.Context) (*http.Response, error)
	getSettings(ctx context.Context) (*http.Response, error)
	updateSettings(ctx context.Context) (*http.Response, error)
	resolveTenant(ctx context.Context) (*http.Response, error)
	getFeatureSnapshot(ctx context.Context) (*http.Response, error)
	setFeature(ctx context.Context) (*http.Response, error)
	superAdminFeatureOverride(ctx context.Context) (*http.Response, error)
}